Day4_Basics_of_RAG/RAG/requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
